initial changeset release rewrite
