+++ scripts/release/utils/tests.ts
@@ -1,0 +1,30 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { spawn } from 'child-process-promise';
+import { projectRoot as root } from '../../utils';
+
+export async function runTests(): Promise<void> {
+  await spawn('yarn', ['test'], {
+    cwd: root,
+    stdio: 'inherit'
+  });
+}
+
+export async function setupTestDeps(): Promise<void> {
+  await spawn('yarn', ['test:setup'], { stdio: 'inherit' });
+}
